+++ vfs/can/vfs-can.c
@@ -127,7 +127,7 @@ static int vfs_can_priv_parse_dts(struct device* pdev, struct hal_can_bus_config
     cfg->can_clk_periph = (uint32_t)can_clk;
     cfg->tx = (struct hal_can_pin_cfg){
         .port = (uintptr_t)tx_port,
-        .pin = (uint16_t)tx_pin,
+        .pin = (uint32_t)tx_pin,
         .clk_bus = (uint32_t)tx_clk,
         .af = (uint32_t)tx_af,
         .output_type = (uint32_t)tx_output_type,
@@ -137,7 +137,7 @@ static int vfs_can_priv_parse_dts(struct device* pdev, struct hal_can_bus_config
     };
     cfg->rx = (struct hal_can_pin_cfg){
         .port = (uintptr_t)rx_port,
-        .pin = (uint16_t)rx_pin,
+        .pin = (uint32_t)rx_pin,
         .clk_bus = (uint32_t)rx_clk,
         .af = (uint32_t)rx_af,
         .output_type = (uint32_t)rx_output_type,
